backend/requirements.txt
# Django e dependências principais
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.3

# Processamento de dados
pandas==2.1.3
numpy==1.24.3
openpyxl==3.1.2
xlrd==2.0.1

# Banco de dados
psycopg2-binary==2.9.9
sqlite3

# Autenticação e segurança
djangorestframework-simplejwt==5.3.0
firebase-admin==6.2.0

# Utilitários
python-decouple==3.8
Pillow==10.1.0
celery==5.3.4
redis==5.0.1

# Desenvolvimento
django-debug-toolbar==4.2.0
django-extensions==3.2.3

# Produção
gunicorn==21.2.0
whitenoise==6.6.0 